所持金UI追加、 UI更新
前提条件矢印が色々おかしい
サイズ、方向、色、アスペクト比
次はこれを修正するつもり

+++ TOC_Game/Assets/Code/C#/DB/BaseUpGrageDB_Gabu.cs
@@ -10,9 +10,9 @@ public class BaseUpGrageDB_Gabu : ScriptableObject
     public string UpGrageName = "Diamond";
     public int UpGrageID = 0;
     [Header("金のコスト")]
-    public BigInteger moneyCost = 0;
+    public long moneyCost = 0;
     [Header("スターのコスト")]
-    public BigInteger starCost = 0;
+    public long starCost = 0;
     [TextArea, Header("詳細情報")]
     public string infometion;
     [Header("画像")]
